Don't display an enormous Twitter icon when CSS is disabled
If CSS is disabled, the SVG for the blue Twitter icon on tweets will render
at full-width, because there are no width/height constraints.  This PR swaps
out the inline data+svg image for one that's loaded from a file at render time
(so it's a bit easier to edit), and adds width/height attributes so it
stays small even when CSS is disabled.