Make /all-posts/ the canonical page for the post archive
While searching in DuckDuckGo, I noticed that it's indexed the per-month
and per-year archive pages, which are really specific views of the overall
archive page.  They shouldn't be surfaced in search, only for people
fiddling with URLs.

This patch uses <link rel="canonical"> to redirect search engines to the
overall archive page, and away from the per-month/per-year pages.