Push generated theme assets into dedicated directories
This means I can put the entire directory in the `keep_files` setting,
rather than wildcards which don't seem to work -- and then we can skip
regenerating the assets on subsequent builds.

This isn't much of an issue for the CSS/favicons, but it's more of a
concern for the header images, which are substantially larger.