Fix the way Twitter avatars get resized
Because `image.resize` returns a new image rather than mutating the
existing image in-place, I was discarding the resize.  Oops!  This fixes
it so avatars do actually get resized.

This lets me delete the "gets bigger when resized" code, because this no
longer occurs when avatars are resized properly.

Most avatars are now <4KB.