Check in all my card assets separately
This is a precursor to removing an old file called `codesamples.key`,
which contained all these cards in a single file.  It was getting very
large and clogging the Git history, because every new card would cause
the entire file to be re-committed.  I'm breaking it into smaller files,
then I'm going to remove all versions of the single file from history.