Use the <time> element for timestamps on posts
According to the MDN documentation, the "24 September 2023" format I'm
using as the child text content isn't a valid date format (all the valid
formats are numeric, so maybe it's to do with internationalisation?).

I suspect search engines already know how to parse this, but it seems
semantically nice to provide the date in a machine-readable format here.