Make the cards generally a lot more efficient
Previously the cards were only available in two sizes: 400px and 800px.
But they're displayed at up to three sizes, depending on the width of the
screen, and 400px isn't a good fit for any of them: it's wider than
necessary for two of them, and not wide enough for the third.

This patch changes the way those cards are generated, so now there are
six sizes (for 1-col, 2-col and 3-col widths, at 1x and 2x each).

This should reduce the size of these pages nicely.