Reduce the CSS at the top of tinted pages
Previously all the default CSS variables were being re-included on any
page that had a custom tint colour, because they were part of `variables.scss`.
Now we don't create any CSS in that file, just export a default vars
mixin that can be used once in the global stylesheet.

This reduces the average page size by ~71 bytes -- not a lot, but
something!