plugins: change how colours are chosen in header images
Previously the set of colours in the mosaic headers was created by converting
the tint colour to HSL, and taking a fixed percentage of the lightness value.
This meant bright colours had a lot of variety, and dark colours had almost
none.

The new approach is converting the tint colour to CIELAB, then varying the
lightness and covering a range of perceptual difference. This gives a more
balanced range of colours and reins in the excess for very light/dark tints.

And it found two bugs in the color gem!