all: stop using the {% highlight %} tag
I was only using this in a handful of posts, and otherwise I prefer to
use {% code %}. The old tag wasn't even working correctly, and it's
breaking the new Python build system. Get rid of it!

While I'm here, I added the new syntax highlighting to some old posts.

For #1138
For #1165