plugins: remove the distinction between index/social cards
I thought I'd want to create separate cards for use on the site index
and on social media, but in practice I don't. These days I use full
bleed images which work in both contexts; the handful of site-specific
cards I created look bad in dark mode.

Ditch them, which makes this code slightly simpler as I port it to
Python for #1165